Fix 1+2+3: thread-safety snapshots, move ContextEvents to objects/core, fix FQN refs
