+++ src/main/kotlin/cz/vutbr/fit/interlockSim/context/Context.kt
@@ -15,7 +15,7 @@ import cz.vutbr.fit.interlockSim.objects.tracks.TrackBlock
 import cz.vutbr.fit.interlockSim.util.ExtendedUnorientedGraph
 import cz.vutbr.fit.interlockSim.util.Point
 import io.github.oshai.kotlinlogging.KotlinLogging
-import cz.vutbr.fit.interlockSim.objects.cells.ContextPropertyChangeListener
+import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
 import org.koin.core.scope.Scope
 
 private val logger = KotlinLogging.logger {}
