+++ src/test/kotlin/cz/vutbr/fit/interlockSim/RaceConditionTest.kt
@@ -37,6 +37,8 @@ import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.atomic.AtomicReference
+import cz.vutbr.fit.interlockSim.objects.core.ContextChangeEvent
+import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
 
 /**
  * Concurrent access tests for railway simulation components.
@@ -796,7 +798,7 @@ class RaceConditionTest : KoinTestBase() {
 						for (iter in 0 until iterations) {
 							// Create a listener
 							val listener =
-								cz.vutbr.fit.interlockSim.objects.cells.ContextPropertyChangeListener { _ ->
+								ContextPropertyChangeListener { _ ->
 									// Listener receives property changes
 								}
 
@@ -986,7 +988,7 @@ class RaceConditionTest : KoinTestBase() {
 			// Arrange
 			val context = buildMinimalSimulation()
 			val threadCount = 3
-			val listenerOps = ArrayList<cz.vutbr.fit.interlockSim.objects.cells.ContextPropertyChangeListener>()
+			val listenerOps = ArrayList<ContextPropertyChangeListener>()
 			val startLatch = CountDownLatch(1)
 			val doneLatch = CountDownLatch(threadCount)
 			val exceptions = CopyOnWriteArrayList<Exception>()
@@ -998,7 +1000,7 @@ class RaceConditionTest : KoinTestBase() {
 					startLatch.await()
 					for (i in 0 until 10) {
 						val listener =
-							cz.vutbr.fit.interlockSim.objects.cells.ContextPropertyChangeListener { _ ->
+							ContextPropertyChangeListener { _ ->
 								// Listener receives events
 							}
 						listenerOps.add(listener)
