+++ src/test/kotlin/cz/vutbr/fit/interlockSim/context/BaseContextTest.kt
@@ -20,6 +20,8 @@ import cz.vutbr.fit.interlockSim.testutil.KoinTestBase
 import cz.vutbr.fit.interlockSim.util.Point
 import org.junit.jupiter.api.Test
 import org.koin.test.inject
+import cz.vutbr.fit.interlockSim.objects.core.ContextChangeEvent
+import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
 
 /**
  * Test suite for BaseContext shared functionality.
@@ -171,7 +173,7 @@ class BaseContextTest : KoinTestBase() {
 			// Arrange
 			var listenerCalled = false
 			val listener =
-				cz.vutbr.fit.interlockSim.objects.cells.ContextPropertyChangeListener { _ ->
+				ContextPropertyChangeListener { _ ->
 					listenerCalled = true
 				}
