+++ src/test/kotlin/cz/vutbr/fit/interlockSim/gui/ModificationTrackerTest.kt
@@ -20,7 +20,7 @@ import cz.vutbr.fit.interlockSim.context.ContextChangeListener
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import cz.vutbr.fit.interlockSim.objects.cells.ContextChangeEvent
+import cz.vutbr.fit.interlockSim.objects.core.ContextChangeEvent
 import java.io.File
 
 /**
