+++ src/test/kotlin/cz/vutbr/fit/interlockSim/integration/ContextLifecycleIntegrationTest.kt
@@ -37,6 +37,8 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import org.koin.test.inject
 import java.io.File
+import cz.vutbr.fit.interlockSim.objects.core.ContextChangeEvent
+import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
 
 /**
  * Integration tests for context lifecycle management.
@@ -319,10 +321,10 @@ class ContextLifecycleIntegrationTest : KoinTestBase() {
 	/**
 	 * Test PropertyChangeListener implementation that captures all events.
 	 */
-	private class TestPropertyChangeListener : cz.vutbr.fit.interlockSim.objects.cells.ContextPropertyChangeListener {
-		val events: MutableList<cz.vutbr.fit.interlockSim.objects.cells.ContextChangeEvent> = mutableListOf()
+	private class TestPropertyChangeListener : ContextPropertyChangeListener {
+		val events: MutableList<ContextChangeEvent> = mutableListOf()
 
-		override fun propertyChange(event: cz.vutbr.fit.interlockSim.objects.cells.ContextChangeEvent) {
+		override fun propertyChange(event: ContextChangeEvent) {
 			events.add(event)
 		}
 	}
